pedsuite: A collection of packages for pedigree analysis in R

This meta-package makes it easy to install and load the ped suite packages for pedigree analysis in R, covering applications in forensic genetics, medical genetics and others. A detailed presentation of the ped suite is given in the book Pedigree Analysis in R (Vigeland, 2021, ISBN: 9780128244302).

Core packages

At the centre of the ped suite we find the package pedtools, on which all the others depend. In addition, several packages contain basic features often needed in many types of analysis. These are the core packages:

pedtools: Creating and working with pedigrees and marker data

verbalisr: Textual descriptions of pedigree relationships

ribd: Pedigree-based relatedness coefficients

pedprobr: Marker probabilities and pedigree likelihoods

forrel: Forensic pedigree analysis and relatedness analysis

Specialised packages

The following packages are devoted to special applications. They must be loaded separately, e.g., library(ibdsim2).

dvir: Disaster victim identification

ibdsim2: Simulation of identity-by-descent sharing by family members

paramlink2: Parametric linkage analysis

pedbuildr: Reconstructing pedigrees from marker data

pedmut: Mutation models for pedigree likelihood computations

segregatr: Segregation analysis for clinical variant interpretation