pedsuite: Easy Installation of the 'pedsuite' Packages for Pedigree Analysis

The 'pedsuite' is a collection of packages for pedigree analysis, covering applications in forensic genetics, medical genetics and more. A detailed presentation of the 'pedsuite' is given in the book 'Pedigree Analysis in R' (Vigeland, 2021, ISBN: 9780128244302).

See also

Author

Maintainer: Magnus Dehli Vigeland m.d.vigeland@medisin.uio.no (ORCID)

Authors:

Thore Egeland thore.egeland@nmbu.no (ORCID)